feat: match more providers (#66)
There were two issues with provider detection in the SDK until now:
- The common providers had some name permutations that we didn't know about
- We didn't add custom names for the providers that do not have default values in semconv, even though semconv allows for custom values. I believe we SHOULD cover every possible provider to the best of our ability.

Therefore, this PR adds:
- openai variants such as openai.embedding
- anthropic.messages
- a bunch of new providers: assemblyai, cerebras, deepgram, deepinfra, elevenlabs, fal, fireworks, gladia, hume, lmnt, luma, replicate, revai, togetherai, vercel

One intentional omission that we are not covering for now is openai-compatible. Need to confirm with the semconv people whether that should be the name, or if it's better to omit it (because it could be any number of providers!)

Mock providers have been updated to get their names detected by the <provider>.<action> case, and tests have been updated accordingly.